+++ backend/teleconsulting/src/main/java/com/had/teleconsulting/teleconsulting/Payloads/PatientDTO.java
@@ -15,7 +15,7 @@ import javax.persistence.Id;
 
 public class PatientDTO {
 
-    int patientID;
+    long patientID;
     String patientFirstName;
     String patientLastName;
     String  patientMobileNumber;
